client: split walk tests
